data/cloud_loaders/requirements.txt
# Cloud Data Loaders Dependencies
python-dotenv>=1.0.0
faker>=28.0.0
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0

# Oracle Cloud
oracledb>=2.0.0

# Databricks
databricks-sql-connector>=3.0.0
databricks-sdk>=0.18.0
